Move wallet related code into 'wallet' folder
