+++ user-registry-ui/ethereum/wallet/actions.test.js
