+++ user-registry-ui/ethereum/wallet/reducer.js
