+++ user-registry-ui/ethereum/wallet/reducer.test.js
