+++ user-registry-ui/ethereum/wallet/selectors.js
