+++ user-registry-ui/ethereum/wallet/selectors.test.js
